desktop-app/backend/requirements.txt
# Dive AI Backend Requirements

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6
sse-starlette>=1.8.0

# HTTP Client
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.26.0

# Config
pyyaml>=6.0

# Desktop Automation
pyautogui>=0.9.54
Pillow>=10.0.0

# Optional: Tesseract for OCR
# pytesseract>=0.3.10

# Optional: Browser automation
# playwright>=1.40.0

# Optional: Cloud Database
# supabase>=2.0.0
